Use modern clipboard API to copy text to the clipboard
It is very convenient to use the clipboard API, instead of the
legacy `document.execCommand('copy')` with all the added pieces.

However, we still need the legacy functionality for old browsers.

Before merging this PR, all supported browsers (in all platforms)
must be tested.

There is a good review of pros and cons of using the clipboard API here: https://stackoverflow.com/a/30810322/14463679