Fix a type checking error related to `getFilterValues`
Fix a typechecking error in `useRootThread`.

`@return` has no effect when used above a variable initialized to a
value that is not an arrow function or function expression. Therefore
the type was inferred based on the `filterValues` local variable to be
`{}`. Correct this to `Object.<string,string>` so that it matches up
with the `ThreadState` type.