Preload `/api/` and `/api/links` responses in boot script
https://github.com/hypothesis/h/pull/5862 introduced an optimization to
startup of the sidebar app by preloading the `/api/` and `/api/links`
requests.

Because this change was made in h it only benefitted the embedded
client, but not the browser extension.

This commit moves the logic to add the `<link rel="preload">` links to
the client's boot script. The upside of this is that it works with all
distributions of the client and can be changed in future without needing
to modify h and the browser extension etc. separately.

The downside is that because the `<link>`
is no longer in the sidebar app's HTML but in the boot script the
prefetch will start a little later.

Since Cloudflare caches the `/api/` and `/api/links` responses, these requests
are still very likely to have completed by the time the client code starts executing.