Sync route before waiting for groups and profile to load
Do not wait for groups and profile to load before calling
`router.sync()` to set the initial route and thereby render the
appropriate content component.

Historically the main view components assumed that the user profile and
groups were already loaded before they were rendered, but this is no
longer the case.

This fixes a brief flash of the sidebar's `TopBar` component inside the
notebook when it initially loads. It also results in the `Annotations`
and `Page Notes` tabs appearing immediately in the sidebar rather than
waiting until the groups list is populated.