Hide the flag icon if user is unauthenticated
Previously, all users were able to click the flag icon. However,
unauthenticated users were shown a toast message if they click the icon:
`You must be logged in to report an annotation`.

The behaviour has been changed so that the flag button is only shown (and clickable) for authenticated users.

Resolves #1037