Do not provide share link if URL is not available on the web
In both the share-single-annotation and share-annotations-on-a-page
components, provide a UI path for when the annotation's URI or the
page's URI (respectively) are not web-available. That is, to be
"shareable", the URI in question needs to have an `http:` or `https:`
protocol.

In cases where the URI is not "shareable," provide some explanatory
text about its non-shareability and don't provide a sharing (bouncer)
link.

This change is intended to avoid confusion when users try to share an
annotation or a set of annotations that have been made on a local
(e.g. PDF) document.

Some SCSS patterns have also been adjusted to accommodate more flexible
styling of the share-single-annotation panel.

Fixes #2786