Refine design and loading state for `NotebookResultCount`
Redesign the results count for the notebook to better handle its
behavior when loading (sometimes large) sets of annotations.

Previously, the annotation count used the `visibleCount` store method,
which returns a count of annotations that are visible on-screen. While
loading, use the service's reported `total` value to show the number of
annotations that will be visible once loading is complete.

Do not show thread count while loading, as this cannot be determined
until all annotations are present and threaded.

Fixes #2864