Adjust UI for sharing annotations on non-shareable documents
In both the share-this-annotation and share-this-page's-annotations
components, provide a UI path for when the annotation's URI or the
page's URI (respectively) are not web-available. That is, to be
"shareable in context", the URI in question needs to have an
`http:` or `https:` protocol.

For "share this page's annotations" when the page is not web-accessible,
explain why the page's annotations can't be shared in context and don't
provide a sharing link, as it won't work.

For "share this annotation," when the annotation's URI (document) isn't
web-accessible, provide some explanatory text about how it can't be
shared in-context, but provide a link to the single-annotation view,
when available.

These changes are intended to avoid confusion when users try to share an
annotation or a set of annotations that have been made on a local
(e.g. PDF) document.

Some SCSS patterns have also been adjusted to accommodate more flexible
styling of the share-single-annotation panel. The width of the compact
panel pattern has been increased slightly so that the new wording
variant for not-shareable-in-context single annotations doesn't risk
taking up too much vertical space.

Fixes #2786