Use router as source of truth for whether app is sidebar or not
Previous commits updated the router service to enable it to correctly
set the route from the initial path in the browser extension. However
some parts of the client still behaved as if they were the sidebar in
the notebook view because there was a second path => route mapping in
`src/sidebar/index.js`.

This commit replaces that with `store.route()` everywhere which is the
route determined by the router service.

This fixes an issue where launching the notebook app in the browser
extension resulted in a "second Discovery server" error in the console.