Remove Jenkins build debugging code
Remove the verbose logging that was enabled for Chrome in Jenkins builds
a while ago. This was used to debug a problem with sporadic failures
that we haven't seen too much recently. If it becomes a significant
problem in future we'll probably take the opportunity to shift the whole
client build to GitHub Actions instead.