Fix incorrect use of `assert.isEmpty()` in UI tests
`assert.isEmpty()` checks the size of an array, map or set or counts the
number of keys on an object. It doesn't know how to check that an Enzyme
wrapper produced no output. `assert.isEmpty(wrapper)` passes due to
implementation details of the Enzyme wrapper - it behaves like an object
with no properties. However it _always_ passes regardless of the number of
nodes in the wrapper, which is not the intended behavior here.