Remove support for passing a non-Element ref to `useElementShouldClose`
`useElementShouldClose` allowed passing a ref to a Preact component
rather than a ref to a DOM element. Support for this relied on Preact
component's having a `base` property, which is going away in future [1]

All of the actual usage of `useElementShouldClose` in the app passed an
Element ref, so I've simply removed the functionality and associated
tests.

[1] https://github.com/preactjs/preact/pull/2971