Move `useElementShouldClose` and `normalizeKeyName` into frontend-shared package
`useElementShouldClose` is a useful generic hook for implementing non-modal dialogs. We
also use it in the LMS frontend, so it makes sense to share the implementation.

This function depends on `normalizeKeyName`, so I moved that as well.

`useElementShouldClose` also depended on a `listen` helper. This helper
is not currently used by any other code so I moved it into
`use-element-should-close.js` as a non-exported helper function and
simplified it.

In the process I found it was necessary to change the gulp task that
builds the frontend-shared/lib/ directory to handle subdirectories under
frontend-shared/src/.