Fix test for `mix-blend-mode: multiply` support in Firefox, recent Chrome
The feature test for the `CSS.supports` API did not work in the most
recent versions of Chrome or in Firefox because `typeof CSS` returns
`object` in those browsers. In Safari and older versions of Chrome it
returns `function`. This change most likely occurred due to
https://chromium-review.googlesource.com/c/chromium/src/+/2434404.

All of the browsers we currently support implement the two-argument
version of `CSS.supports` (see [1]) and so we can fix the issue by removing the feature test.

Note that the single-argument version of `CSS.supports` is not supported
by some of the oldest browsers that we support.

[1] https://developer.mozilla.org/en-US/docs/Web/API/CSS/supports#browser_compatibility