Add a UI component demo app
Add a new `/ui-playground` route to the client's dev server which hosts
a UI component playground/demo app, plus a link to this route on the dev
server homepage. The demo app provides a place to see and experiment with
UI components outside the context of the sidebar/notebook apps.

 - Add new `/ui-playground` route to dev server which serves a small
   Preact SPA

 - Add Preact SPA in `dev-server/ui-playground/` with a couple of
   placeholder "demos" for the `Button` and `Menu` components

 - Add configuration to gulpfile.js to build ui-playground JS and CSS
   bundles for Preact SPA