Update annotator toolbar CSS class names to match conventions
Rename CSS class names for `Toolbar` component to match the component itself.
This component name is really overly generic but at least this brings
the CSS and Preact components into line.

One of the toolbar's classes was applied to a container element created
by `ToolbarController`. This class has been moved to an element created
by the `Toolbar` component for consistency. This means there is an
apparently redundant (no styling, one child) container div when looking
at the DOM structure for the toolbar + bucket bar. This will go away if
the whole sidebar container is made into a single Preact component in
future.