Align bucket bar class names with current conventions
Rename CSS classes used by the bucket bar to match the `Buckets`
component name and move the corresponding styles into
src/styles/annotator/components/Buckets.scss.

As was recently done with the sidebar's vertical toolbar, the styles on the bucket bar
container were moved to an element rendered by the `Buckets` component
itself, so that there is a 1:1 association between JS modules and SASS
modules for the bucket bar.