Improve filter options list in notebook
If the current user shows up in the filter list. (e.g they have annotations), then sort that user to the front of the list and add " (Me)" suffix to the end of the displayed name to make it easier for a user to find and filter their own annotations.