Fix PDF side-by-side mode not activating
The `ToolbarController#getWidth` method started incorrectly returning 0
after 243c06d3d01172e25dfc86c823d0c5b3dda04e53. This commit moved the
content of the toolbar to a child of `ToolbarController._container` and
the `getWidth` method was not updated accordingly.

The incorrect result of `this.toolbar.getWidth()` in `Sidebar` later
resulted in an incorrect calculation of the sidebar's width and caused
`PdfSidebar` to think that the window was too narrow to support
side-by-side mode.

Fix the problem by updating `ToolbarController#getWidth` so that it
returns the correct value again.