Ensure PDF integration is loaded in PDF documents
The recent refactor to change the relationship between the `Guest` and `Sidebar`
classes broke PDF.js integration because the `PdfSidebar` class is no
longer able to modify the config that is passed to the `Guest` class.
Previously it would pass `PDF: {}` in the config to the `super()` call
which would then result in the `Guest` class loading the PDF
integration.

This now has to be done in the `index.js` module. We already had to do
this previously for ebooks. Now it has to be done for all scenarios.