Refactor Guest#createAnnotation for readability
Convert the `createAnnotation` method of `Guest` to use async/await
instead of manual chaining of Promise operations. This makes it easier
to follow the logic and see that document metadata is assigned to the
annotation before it is broadcast to other parts of the annotator in the
`beforeAnnotationCreated` event. It also makes it easier to external
code (eg. the tests) to determine when `createAnnotation` is "done" and
has finished updating the annotation.

 - Convert `Guest#createAnnotation` to use async/await

 - Update tests for `createAnnotation` to await the returned Promise
   instead of a fixed timeout

 - Add a missing check in `createAnnotation` tests for the argument
   passed to the `beforeAnnotationCreated` event