Render bucket bar immediately upon construction
Render the bucket bar immediately using the Guest's current set of anchors when it
is constructed.

This resolves an issue introduced in 68d44395 that the background of the
bucket bar did not appear until after the first batch of annotations had
been fetched from the API and synced from the sidebar to the annotator.
It also ensures that buckets are drawn for any anchors that already
exist in the Guest when the BucketBar is initialized.