Change `mockImportedComponents` to not rely on existence of `propTypes`
Detecting a Preact/React functional component that does not use `propTypes` is
impossible to do with 100% accuracy. `function Foo() { return 'Hello' }`
is a perfectly valid component with no obvious Preact/React features.

Fortunately because of the way this helper is used, we can make do with an imperfect
check. If `mockImportedComponents` mocks something which is not really a
component, that mock will be overridden with a real mock later.