Simplify "no selectable text" banner styling
Avoid the need to customize various inline styles of the
`<hypothesis-banner>` element by positioning it at the top of the DOM as
the first child of `document.body`. As a result it naturally pushes the
PDF.js UI down. The only change we need to make to PDF.js's UI is to adjust
the size of the `#outerContainer` element so that it does not overflow the
bottom of the viewport.

Aside from making styling simpler, making the banner's DOM position
better match its visual position may help with reading order for screen readers.
Related to this the banner is currently missing useful semantic
information for a11y purposes.

I also removed the banner resizing handling for the moment on the basis that it
isn't that important.