Fetch annotations in parallel with groups and profile where possible
Currently both the sidebar and notebook wait until the user's profile and groups are
fetched before fetching annotations. This is because the search query to
fetch annotations depends on the focused group. The groups query may
take several hundred ms to complete, so this ends up delaying the
initial fetch and display of annotations by that amount of time.

In the notebook the common case is that the user is already logged in
and the notebook is configured to display a specific group ("the direct
linked group"): whatever the user had previously selected in the sidebar.
In this scenario it is very likely that the direct-linked group will end
up being the focused one after groups are fetched.

Based on these assumptions, we can speed up the initial annotation by
loading annotations from the direct-linked group. If the focused group
ends up being different, annotations will just be re-fetched from the
correct group once that is known.