Optimize perceived loading time by varying /api/search page size
When fetching annotations from the backend there is a trade-off between
the time taken to generate a page (less if the page size is smaller) and
the time taken to fetch all pages (less if the page size is larger).

To optimize the perceived loading time we can pick a smaller page size
for the first page, enabling the first screenful of annotations/threads
to appear quickly, and then use a larger page size for the remaining
pages.

 - Change the `pageSize` option for `SearchClient` from a fixed number
   to a callback that returns the page size for a given page index

 - Set the default value for the `pageSize` option to use a small page
   size (50) for the first page and a larger size (200) for remaining
   pages