Increased number of threads per page in notebook
We have done a little investigation on the number of threads that are
visible on a pages in notebook:
- on mobile: up to 4 threads
- on tables: up to 7 threads
- on mobile: up to 8 threads

These numbers are just rough estimates since zoom level and screen
orientation affects these numbers.

Our aim: we want to display the larger number of threads without forcing
users to do excessive scrolling.

We are going to test with 25 threads per page and see what are users
think about that.