Remove Google Analytics from the client
We have had Google Analytics disabled for the Hypothesis client for over
a year and do not plan to re-enable it. We may need analytics for other
purposes in future (for example, to help inform product design). In that
case we will likely implement a solution that is a better fit for
whatever those needs end up being.

This commit removes the existing analytics code in the meantime so that
we don't spend time maintaining something that is not used.

Fixes #3027