Extract side-by-side functionality out of PdfSidebar class
Move the functionality that PdfSidebar implements elsewhere and remove
the class. This mainly consists of side-by-side mode, but there is also
the logic to get the scrollable document container for use with the
bucket bar. The PDF-specific logic has been moved to the `PDF` class in
`src/annotator/plugin/pdf.js` and the generic logic has been moved to
the `Sidebar` or `Guest` classes.

This change is a step towards creating an common interface between the
Guest/Sidebar and the document-type/viewer specific functionality. In
future the `PDF` class will evolve into the implementation of this
interface for PDFs. For HTML documents the logic is currently contained
directly in the `Guest` class and this will be extracted out into a
class that implements the same interface.

 - Move the PDF-specific parts of side-by-side mode into
   `PDF#fitSideBySide`
 - Move the generic parts of side-by-side mode into the `Sidebar` class.
   This will allow it to be reused for HTML documents or other document
   types in future
 - Move the logic for determining the scrollable container into
   `PDF#contentContainer`
 - Change the annotator setup code in `annotator/index.js` to always
   instantiate the `Sidebar` class, regardless of document type