Change thread sorter "Newest" & "Oldest" to use created
This changes the sorting in the notebook to sort by date created rather than last edited date ("updated"). This prevents root threads from resorting to the top after they are edited in the notebook.  This also changes the way the sidebar orders root threads when using Newest/Oldest sorting filter in the same manner.