Hide sidebar on 'mousedown' event
When the selection of text finishes on the sidebar's iframe, the click
event is not fired. Hence, the sidebar is not collapsed and the adder is
not displayed. This is because events don't bubble up through the
iframe.

This PR makes the sidebar to collapse on 'mousedown' events on the host
page. This is similar in nature to the 'touchstart' event that guest
already listens (although for a different reason).

Closes #463