Add shared buttons components
Add shared button components in a temporary "shared/components"
directory. The plan is to extract this to the `frontend-shared` package
once it has been more further integrated into the client preact app(s)
and the kinks are worked out. This will replace
`src/sidebar/components/Button` and will be reusable in other apps