Implement side-by-size mode for HTML documents
In PDF documents we show the sidebar alongside the content when there is enough
horizontal space available. Extend this behavior to HTML documents.

In HTML documents the content can be divided into the _main content_ and
_peripheral content_. The main content consists of the text that the user is
likely to be annotating (eg. the body of the news or encyclopedia article), and
the peripheral content refers to navbars, ads etc. on the side of the page. In
side-by-side mode we want to maxize the horizontal space available for the main
content, but can afford to allow the sidebar to cover peripheral content.

Determining what part of the HTML document consistutes the main content is done
by heuristics implemented in `guessMainContentArea`. These heuristics work by
looking for common vertical margins among the largest paragraphs of text on the
page. If the heuristics are unable to determine what the main content is,
side-by-side mode is not used and the client behaves the same as before.

Resizing the document when side-by-side mode is active is done by first
shrinking the document's body element by the full width of the sidebar when it
opens and then increasing the body's width if there is space between the main
content and the edge of the sidebar. Any peripheral content in this space will
be covered by the sidebar.

Resizing the document will generally cause the content to shift. To try and
ensure the same content remains visible before and after opening the sidebar, a
`preserveScrollPosition` helper attempts to determine how much the content
shifted after the relayout and adjust the scroll position to keep, near-enough,
the same content visible. To do this, it chooses an element near the top of the
screen as an anchor and checks how much it shifts following the resize. It then
adjusts the document's scroll position to compensate.