Remove `plugin` directory from src/annotator/
The term "plugin" is vestigial. The remaining cross-frame.js module in this
directory has been moved up to the top level of `src/annotator/` since
there isn't a more specific home that is suitable. Also change the
`CrossFrame` class to be a named rather than default export per current
conventions.