Refactor config/index.js and add namespaces
In an effort to filter the config based on the application type (sidebar, notebook, etc), there were some aspects on the config and settings that were confusing. This refactor moves all the rules to a single configuration definition that makes it clear 1. where a setting comes from and 2. what rules it follows before getting passed along to components.

Additionally, it was unclear which settings were allowed in the browser extension context because only the hostPageSetting settings paid attention
to the flag and some of the getters did not.

Changes:
- When in browser context, values which are prohibited are now omitted rather than set to null
corers methods
- When requeuing a config using a namespace, values outside of that namespace are omitted.

Note: Currently, all the namespaces request all values, but that will need to change.

TODO:
- Remove custom coerce code form some of the getters in settings.js
- Remove allowInBrowserExt flag in settings.js
- Modify namespaces as needed for "annotator" and "sidebar"
- Can contentContainer be moved into settings?
- Use HostConfig typedef and use for Sidebar()
- Add AnnotatorConfig typedef and use for Guest()
- Review / add any other config types needed for notebook, pdf, etc.