Refactor config/index.js to add contexts
Refactor config/index.js in an effort to filter the config based on the application context (sidebar, notebook, annotator, This refactor moves all the rules to a single configuration definition that makes it clear 1. where a setting comes from and 2. what rules it follows before getting passed along to the rest of the app.

Additionally, it was unclear which settings were allowed in the browser extension context because only the hostPageSetting settings made use of the boolean flag but some of the getters did not.

Changes:
- When in browser context (isBrowserExtension), values which are prohibited are now omitted rather than set to null
- When requeuing a config using an app context, values outside of that context are omitted.