Show a clearer error if `window.open` call fails
If opening the login popup window using `window.open` fails or succeeds
but fails to return a `Window` reference to the caller then show a more
helpful error message.

There are various rare scenarios (browser bugs, browser extensions)
which might cause this to happen. There is not a whole lot we can do to
recover, but at least we can show a better error message.