Remove fallback document URL and metadata
Remove the fallback logic that uses `location.href` and `document.title`
as the document URL and metadata if using the integration-specific code
(HTML or PDF) to get the document URL and metadata fails. Instead just
let the async exception propagate.  Currently this will result in no
annotation/highlight being created and a console error. In future we
should improve this by showing an error notification to the user.

Recording the wrong document metadata with an annotation is worse than
not creating an annotation at all: The user will be unaware that there
is a problem yet the annotation will likely fail to show up in future or
be displayed with completely wrong metadata in the Notebook. Therefore
we want the failure here to be obvious so that the user notifies us of
the problem and doesn't create a bunch of annotations which later
"disappear".

Fixes #3204