Convert existing `window.confirm` calls to use `confirm` prompt
Convert existing uses of `window.confirm` to use our own `confirm`
utility which uses the standard Dialog from our pattern library.

Using our own prompt avoids an immediate problem that browsers, starting
with Chrome v91, are starting to block the use of `window.confirm` in
third-party iframe. It also makes the visual design of the prompt match
the rest of the application.